created modules directory and moved the weather and movies functionalities into their respective files in modules
